perubahan

+++ app/Http/Controllers/KabupatenController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Models\Kabupaten;
 use App\Models\Provinsi;
 use RealRashid\SweetAlert\Facades\Alert;
+use Illuminate\Validation\Rule;
 
 class KabupatenController extends Controller {
 
@@ -22,10 +23,11 @@ class KabupatenController extends Controller {
 
     public function store( Request $request ) {
         $validated = $request->validate( [
-            'nama' => 'required|string|max:255',
+            'nama' => 'required|string|max:255|unique',
             'provinsi_id' => 'required|integer|exists:provinsi,id',
         ], [
             'nama.required' => 'Nama Kabupaten wajib diisi.',
+            'nama.unique' => 'Nama Kabupaten sudah terdaftar.',
             'provinsi_id.required' => 'Nama Provinsi wajib diisi.',
         ] );
 
@@ -44,15 +46,21 @@ class KabupatenController extends Controller {
     }
 
     public function update( Request $request, $id ) {
-        $validated = $request->validate( [
-            'nama' => 'required|string|max:255',
+
+        $kabupaten = Kabupaten::findOrFail( $id );
+        $data = $request->validate( [
+            'nama' => ['required', 'string', 'max:255',
+                Rule::unique('kabupaten', 'nama')->ignore($kabupaten)
+        ],
+            'provinsi_id' => 'required|integer|exists:provinsi,id'
         ], [
             'nama.required' => 'Nama Kabupaten wajib diisi.',
+            'nama.unique' => 'Kabupaten ini telah terdaftar silahkan masukan kabupaten lain nya.',
+            'provinsi_id.required' => 'Provinsi wajib diisi.'
         ] );
 
         try {
-            $kabupaten = Kabupaten::finOrFail( $id );
-            $kabupaten->update ( $request->all() );
+            $kabupaten->update($data);
         } catch ( \Exception $e ) {
             Alert::error( 'Gagal', 'Terjadi kesalahan saat menyimpan data: ' . $e->getMessage() );
         }
@@ -60,6 +68,10 @@ class KabupatenController extends Controller {
     }
 
     public function destroy( $id ) {
+        $kabupaten = Kabupaten::findOrFail( $id );
+        if ($kabupaten->kecamatans()->exists()) {
+            return redirect('/kabupaten')->with('error', 'Kabupaten tak bisa di hapus karna memiliki kecamatan yang terkait.');
+        }
         try {
             $kabupaten = Kabupaten::findOrFail( $id );
             $kabupaten->delete();
